Upgrading patch level gems
Fixing Nokogiri's security vulnerability.

This app, however, is broken because google finally retired their image
search public API.
Need to register the app as a third party search for google and obtain a
valid API token for Flickr. Will have to wait for the holidays. Created
issues #4 and #5 to handle both.